requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.7
# by the following command:
#
#    pip-compile requirements.in
#
absl-py==2.1.0
    # via
    #   tensorboard
    #   tensorflow-intel
astunparse==1.6.3
    # via tensorflow-intel
cachetools==5.5.2
    # via google-auth
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
cycler==0.11.0
    # via matplotlib
flatbuffers==25.2.10
    # via tensorflow-intel
fonttools==4.38.0
    # via matplotlib
gast==0.4.0
    # via tensorflow-intel
google-auth==2.38.0
    # via
    #   google-auth-oauthlib
    #   tensorboard
google-auth-oauthlib==0.4.6
    # via tensorboard
google-pasta==0.2.0
    # via tensorflow-intel
grpcio==1.62.3
    # via
    #   tensorboard
    #   tensorflow-intel
h5py==3.8.0
    # via tensorflow-intel
idna==3.10
    # via requests
importlib-metadata==6.7.0
    # via markdown
joblib==1.3.2
    # via scikit-learn
keras==2.11.0
    # via tensorflow-intel
kiwisolver==1.4.5
    # via matplotlib
libclang==18.1.1
    # via tensorflow-intel
markdown==3.4.4
    # via tensorboard
markupsafe==2.1.5
    # via werkzeug
matplotlib==3.5.3
    # via -r requirements.in
numpy==1.21.6
    # via
    #   -r requirements.in
    #   h5py
    #   matplotlib
    #   opt-einsum
    #   scikit-learn
    #   scipy
    #   tensorboard
    #   tensorflow-intel
oauthlib==3.2.2
    # via requests-oauthlib
opt-einsum==3.3.0
    # via tensorflow-intel
packaging==24.0
    # via
    #   matplotlib
    #   tensorflow-intel
pillow==9.5.0
    # via matplotlib
protobuf==3.19.6
    # via
    #   tensorboard
    #   tensorflow-intel
pyasn1==0.5.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pyparsing==3.1.4
    # via matplotlib
python-dateutil==2.9.0.post0
    # via matplotlib
pyyaml==6.0.1
    # via -r requirements.in
requests==2.31.0
    # via
    #   requests-oauthlib
    #   tensorboard
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
scikit-learn==1.0.2
    # via -r requirements.in
scipy==1.7.3
    # via scikit-learn
six==1.17.0
    # via
    #   astunparse
    #   google-pasta
    #   python-dateutil
    #   tensorflow-intel
tensorboard==2.11.2
    # via tensorflow-intel
tensorboard-data-server==0.6.1
    # via tensorboard
tensorboard-plugin-wit==1.8.1
    # via tensorboard
tensorflow==2.11.0
    # via -r requirements.in
tensorflow-estimator==2.11.0
    # via tensorflow-intel
tensorflow-intel==2.11.0
    # via tensorflow
tensorflow-io-gcs-filesystem==0.31.0
    # via tensorflow-intel
termcolor==2.3.0
    # via tensorflow-intel
threadpoolctl==3.1.0
    # via scikit-learn
typing-extensions==4.7.1
    # via
    #   importlib-metadata
    #   kiwisolver
    #   tensorflow-intel
urllib3==2.0.7
    # via requests
werkzeug==2.2.3
    # via tensorboard
wheel==0.42.0
    # via
    #   astunparse
    #   tensorboard
wrapt==1.16.0
    # via tensorflow-intel
zipp==3.15.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
